Added buoyancy script
Instructions:
Add the script to a body of water (a cube) and attach a box that you want to float in the water to the transform. Damper is what fakes friction to slow the box over time, setting it to 1 is a good starting point but feel free to play with it. Box density is used to calculate the mass of the object and is very important to know whether an object will float or sink (0.4 is wood while 2.7 is aluminum). I made the fluid water (density 1.0), to have an object float it must be below water's density and to sink above water's density.